Create decrypt_payload.py
to decrypt the encrypted exfiltrated data received from the infected Android device. for analyzing the exfiltrated data. (note: code on kali /attacker machine )  as we cannot read directly .bin file, as it is encrypted. 


To decrypt the data inside the payload. 
Bin file, we have written a small Python script that will read the file which contains AES
encrypted BASE64 data and decrypt it using the same AES key and Initialization Vector 
from the flutter app.